requirements.txt
flask
twilio
openai
google-api-python-client
google-auth
python-dotenv
